// src/pages/vou/intermediary-sale-order/v2/types.ts
export type IntermediaryContainerType = 'NONE' | 'SOLVENT' | 'RESIN'
// ...
export interface IntermediaryContainerBalance {
  solvent: number
  resin: number
}
// ...
export interface IntermediarySignoffLineView {
  rootLineId: string
  signedQuantity: string
  rejectedQuantity: string
  lossQuantity: string
  remark?: string
}